refactor: uiModel, domainModel 프로퍼티 타입 Int -> Long 수정

+++ android/app/src/main/java/com/created/team201/presentation/home/adapter/DashboardAdapter.kt
@@ -17,7 +17,7 @@ class DashboardAdapter(
     }
 
     override fun getItemId(position: Int): Long {
-        return getItem(position).studyId.toLong()
+        return getItem(position).studyId
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): DashboardViewHolder {
